Server/requirements.txt
Django==1.11
confusable-homoglyphs==3.0.0
django-registration==2.3
django-countries==5.0
django-debug-toolbar==1.8
django-material==1
django-cachalot==1.5.0
python-memcached==1.58
django-rosetta==0.7.13
django-cookie-law==1.0.13
djangorestframework
